__backend_2/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
Django==3.2
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
idna==3.4
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
Pillow==10.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.3
django-cors-headers==3.13.0
gunicorn==20.1.0